+++ Source/ColonyManagerRedux/Comps/CompManagerStation.cs
@@ -7,4 +7,16 @@ namespace ColonyManagerRedux;
 public class CompManagerStation : ThingComp
 {
     public CompProperties_ManagerStation Props => (CompProperties_ManagerStation)props;
+
+    public override IEnumerable<Gizmo> CompGetGizmosExtra()
+    {
+        yield return new Command_Action
+        {
+            action = () => Find.MainTabsRoot.SetCurrentTab(
+                ManagerMainButtonDefOf.ColonyManagerRedux_Manager),
+            defaultLabel = "ColonyManagerRedux.ManagerStation.OpenManagerTab".Translate(),
+            defaultDesc = "ColonyManagerRedux.ManagerStation.OpenManagerTab.Tip".Translate(),
+            icon = Resources.ManagerTab_Gizmo,
+        };
+    }
 }