+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Colony Manager Redux research tab now has a title and description.
 -   Show an alert if jobs aren't being updated in a reasonable amount of time.
+-   Added a gizmo to the manager buildings that takes you to the manager tab.
 
 ### Changed